+++ Controllers/EventController.cs
@@ -1,22 +1,19 @@
-﻿using CamplusBetaBackend.Data.Entities;
-using CamplusBetaBackend.Data;
-using CamplusBetaBackend.Models;
+﻿using CamplusBetaBackend.Models;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using CamplusBetaBackend.Services.Interfaces;
-using System.Runtime.CompilerServices;
 using Amazon.S3.Model;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Amazon;
 using Amazon.S3;
 using Amazon.Runtime;
 using System.Net;
-using CamplusBetaBackend.Services.Implentations;
+using CamplusBetaBackend.DTOs;
 
 namespace CamplusBetaBackend.Controllers {
     [ApiController]
     [Route("api/[controller]")]
     public class EventController : Controller {
+        private readonly string? ADMIN_API_KEY = Environment.GetEnvironmentVariable("ADMIN_API_KEY");
+        private readonly string? RO_API_KEY = Environment.GetEnvironmentVariable("RO_API_KEY");
         private readonly ILogger<EventController> _logger;
         private readonly IEventService _eventService;
         private readonly IAmazonS3 _s3Client;
@@ -32,40 +29,64 @@ namespace CamplusBetaBackend.Controllers {
         }
 
         [HttpGet("GetEvents")]
-        public async Task<ActionResult<Event[]>> GetEvents() {
+        public async Task<ActionResult<Event[]>> GetEvents([FromHeader(Name = "X-API-Key")] string apiKey) {
+        //public async Task<ActionResult<Event[]>> GetEvents() {
+            if (!IsApiKeyValid(apiKey, false)) {
+                return Unauthorized("Invalid API Key");
+            }
+
             Event[]? events = await _eventService.GetEventsFromDB();
 
             if (events == null) {
-                return NotFound();
+                return BadRequest("There is no events.");
             }
             return Ok(events);
         }
 
         [HttpPost("AddNewEvent")]
-        public async Task<ActionResult> AddNewEvent([FromBody] Event @event) {
-            try {
-                @event.EventId = Guid.NewGuid();
-                await _eventService.AddNewEventToDB(@event);
-                return Ok();
-            } catch (Exception e) {
-                return StatusCode(500, "Internal server error: " + e.Message);
+        public async Task<ActionResult> AddNewEvent([FromBody] EventDTO eventDTO, [FromHeader(Name = "X-API-Key")] string apiKey) {
+            if (!IsApiKeyValid(apiKey, true)) {
+                return Unauthorized("Invalid API Key");
             }
+
+            Event newEvent = new Event {
+                EventId = Guid.NewGuid(),
+                Title = eventDTO.Title,
+                StartDateTime = eventDTO.StartDateTime,
+                EndDateTime = eventDTO.EndDateTime,
+                Location = eventDTO.Location,
+                Link = eventDTO.Link,
+                ImageLink = eventDTO.ImageLink,
+                HostId = eventDTO.HostId,
+                ClubId = eventDTO.ClubId,
+                Description = eventDTO.Description,
+            };
+            await _eventService.AddNewEventToDB(newEvent);
+            return Ok(newEvent);
         }
 
         [HttpDelete("DeleteEvent")]
-        public async Task<ActionResult> DeleteEvent(Guid id) {
+        public async Task<ActionResult> DeleteEvent(Guid id, [FromHeader(Name = "X-API-Key")] string apiKey) {
+            if (!IsApiKeyValid(apiKey, true)) {
+                return Unauthorized("Invalid API Key");
+            }
+
             Event? response = await _eventService.DeleteEventFromDB(id);
 
             if (response == null) {
-                return NotFound("Event not found.");
+                return BadRequest("Event not found.");
             }
             return Ok(response);
         }
 
         [HttpPost("UploadEventImage")]
-        public async Task<ActionResult> UploadEventImage(IFormFile image) {
+        public async Task<ActionResult> UploadEventImage(IFormFile image, [FromHeader(Name = "X-API-Key")] string apiKey) {
+            if (!IsApiKeyValid(apiKey, true)) {
+                return Unauthorized("Invalid API Key");
+            }
+
             if (image == null || image.Length == 0) {
-                return StatusCode(500, "Internal server error: Image file is empty");
+                return BadRequest("Internal server error: Image file is empty");
             }
 
             string fileName = Guid.NewGuid().ToString() + Path.GetExtension(image.FileName);
@@ -93,7 +114,11 @@ namespace CamplusBetaBackend.Controllers {
         }
 
         [HttpDelete("DeleteEventImage")]
-        public async Task<ActionResult> DeleteEventImage(string imageLink) {
+        public async Task<ActionResult> DeleteEventImage(string imageLink, [FromHeader(Name = "X-API-Key")] string apiKey) {
+            if (!IsApiKeyValid(apiKey, true)) {
+                return Unauthorized("Invalid API Key");
+            }
+
             Uri uri = new Uri(imageLink);
             string fileName = Path.GetFileName(uri.AbsolutePath);
 
@@ -107,6 +132,15 @@ namespace CamplusBetaBackend.Controllers {
                 return Ok();
             } catch (AmazonS3Exception e) {
                 return BadRequest($"Internal server error: {e}");
+            }
+        }
+
+        private bool IsApiKeyValid(string apiKey, bool isAdmin) {
+            if (isAdmin) {
+                return apiKey == ADMIN_API_KEY;
+            }
+            else {
+                return apiKey == ADMIN_API_KEY || apiKey == RO_API_KEY;
             }
         }
     }
