+++ Controllers/HostController.cs
@@ -1,12 +1,14 @@
-﻿using CamplusBetaBackend.Models;
+﻿using CamplusBetaBackend.Services.Interfaces;
+using CamplusBetaBackend.DTOs;
+using CamplusBetaBackend.Models;
 using Microsoft.AspNetCore.Mvc;
-using CamplusBetaBackend.Services.Interfaces;
-using CamplusBetaBackend.Services.Implentations;
 
 namespace CamplusBetaBackend.Controllers {
     [ApiController]
     [Route("api/[controller]")]
     public class HostController : Controller {
+        private readonly string? ADMIN_API_KEY = Environment.GetEnvironmentVariable("ADMIN_API_KEY");
+        private readonly string? RO_API_KEY = Environment.GetEnvironmentVariable("RO_API_KEY");
         private readonly ILogger<HostController> _logger;
         private readonly IHostService _hostService;
 
@@ -16,45 +18,74 @@ namespace CamplusBetaBackend.Controllers {
         }
 
         [HttpGet("GetHosts")]
-        public async Task<ActionResult<Models.Host[]>> GetHosts() {
+        public async Task<ActionResult<Models.Host[]>> GetHosts([FromHeader(Name = "X-API-Key")] string apiKey) {
+            if (!IsApiKeyValid(apiKey, false)) {
+                return Unauthorized("Invalid API Key");
+            }
+
             Models.Host[]? hosts = await _hostService.GetHostsFromDB();
 
             if (hosts == null) {
-                return NotFound();
+                return BadRequest("There is no hosts.");
             }
             return Ok(hosts);
         }
 
         [HttpGet("GetHost/{id}")]
-        public async Task<ActionResult<Models.Host>> GetHost(Guid id) {
+        public async Task<ActionResult<Models.Host>> GetHost(Guid id, [FromHeader(Name = "X-API-Key")] string apiKey) {
+            if (!IsApiKeyValid(apiKey, false)) {
+                return Unauthorized("Invalid API Key");
+            }
+
             Models.Host? host = await _hostService.GetHostFromDB(id);
 
             if (host == null) {
-                return NotFound();
+                return BadRequest("Could not find host.");
             }
             return Ok(host);
         }
 
         [HttpPost("AddNewHost")]
-        public async Task<ActionResult> AddNewHost([FromBody] Models.Host host) {
-            try {
-                host.HostId = Guid.NewGuid();
-                await _hostService.AddNewHostToDB(host);
-                return Ok();
+        public async Task<ActionResult> AddNewHost([FromBody] HostDTO hostDTO, [FromHeader(Name = "X-API-Key")] string apiKey) {
+            if (!IsApiKeyValid(apiKey, true)) {
+                return Unauthorized("Invalid API Key");
             }
-            catch (Exception e) {
-                return StatusCode(500, "Internal server error: " + e.Message);
+
+            Models.Host? existingHost = await _hostService.GetHostFromDB(hostDTO.Name);
+        
+            if (existingHost == null) {
+                Models.Host newHost = new Models.Host {
+                    HostId = Guid.NewGuid(),
+                    Name = hostDTO.Name,
+                };
+                await _hostService.AddNewHostToDB(newHost);
+                return Ok(newHost);
+            } else {
+                return BadRequest($"Host {hostDTO.Name} already exists.");
             }
         }
 
         [HttpDelete("DeleteHost")]
-        public async Task<ActionResult> DeleteHost(Guid id) {
+        public async Task<ActionResult> DeleteHost(Guid id, [FromHeader(Name = "X-API-Key")] string apiKey) {
+            if (!IsApiKeyValid(apiKey, true)) {
+                return Unauthorized("Invalid API Key");
+            }
+
             Models.Host? response = await _hostService.DeleteHostFromDB(id);
 
             if (response == null) {
-                return NotFound("Host not found.");
+                return BadRequest("Host not found.");
             }
             return Ok(response);
+        }
+
+        private bool IsApiKeyValid(string apiKey, bool isAdmin) {
+            if (isAdmin) {
+                return apiKey == ADMIN_API_KEY;
+            }
+            else {
+                return apiKey == ADMIN_API_KEY || apiKey == RO_API_KEY;
+            }
         }
     }
 }
